+++ internal/service/user_service.go
@@ -18,20 +18,17 @@ func NewUserService(userRepo repository.UserRepository) *UserService {
 	}
 }
 
-// GetAllUsers - получить всех пользователей
 func (s *UserService) GetAllUsers() ([]models.User, error) {
 	return s.userRepo.GetAll()
 }
 
-// GetUserByID - получить пользователя по ID
-func (s *UserService) GetUserByID(id int) (*models.User, error) {
+func (s *UserService) GetUserByID(id int64) (*models.User, error) { // int64
 	if id <= 0 {
 		return nil, errors.New("неверный ID пользователя")
 	}
 	return s.userRepo.GetByID(id)
 }
 
-// GetUserByUsername - получить пользователя по имени (для авторизации)
 func (s *UserService) GetUserByUsername(username string) (*models.User, error) {
 	if username == "" {
 		return nil, errors.New("логин не может быть пустым")
@@ -39,9 +36,7 @@ func (s *UserService) GetUserByUsername(username string) (*models.User, error) {
 	return s.userRepo.GetByUsername(username)
 }
 
-// CreateUser - создание пользователя
 func (s *UserService) CreateUser(username, email, password, role string) (*models.User, error) {
-	// Валидация
 	if username == "" {
 		return nil, errors.New("логин обязателен")
 	}
@@ -58,7 +53,6 @@ func (s *UserService) CreateUser(username, email, password, role string) (*model
 		return nil, errors.New("роль должна быть 'admin' или 'user'")
 	}
 
-	// Хэшируем пароль
 	hashedPassword, err := auth.HashPassword(password)
 	if err != nil {
 		return nil, errors.New("ошибка хэширования пароля")
@@ -79,8 +73,7 @@ func (s *UserService) CreateUser(username, email, password, role string) (*model
 	return user, nil
 }
 
-// UpdateUser - обновление пользователя
-func (s *UserService) UpdateUser(id int, username, email, role, newPassword string) error {
+func (s *UserService) UpdateUser(id int64, username, email, role, newPassword string) error { // int64
 	if id <= 0 {
 		return errors.New("неверный ID пользователя")
 	}
@@ -94,7 +87,6 @@ func (s *UserService) UpdateUser(id int, username, email, role, newPassword stri
 		return errors.New("роль должна быть 'admin' или 'user'")
 	}
 
-	// Получаем текущего пользователя
 	user, err := s.userRepo.GetByID(id)
 	if err != nil {
 		return err
@@ -104,7 +96,6 @@ func (s *UserService) UpdateUser(id int, username, email, role, newPassword stri
 	user.Email = email
 	user.Role = role
 
-	// Если передан новый пароль - хэшируем
 	if newPassword != "" {
 		if len(newPassword) < 6 {
 			return errors.New("пароль должен быть не менее 6 символов")
@@ -115,25 +106,22 @@ func (s *UserService) UpdateUser(id int, username, email, role, newPassword stri
 		}
 		user.PasswordHash = hashedPassword
 	} else {
-		user.PasswordHash = "" // пустая строка = не менять пароль
+		user.PasswordHash = ""
 	}
 
 	return s.userRepo.Update(user)
 }
 
-// DeleteUser - удаление пользователя (с проверкой последнего админа)
-func (s *UserService) DeleteUser(id int) error {
+func (s *UserService) DeleteUser(id int64) error { // int64
 	if id <= 0 {
 		return errors.New("неверный ID пользователя")
 	}
 
-	// Получаем пользователя
 	user, err := s.userRepo.GetByID(id)
 	if err != nil {
 		return err
 	}
 
-	// Если это админ - проверяем, не последний ли он
 	if user.Role == "admin" {
 		adminCount, err := s.userRepo.CountAdmins()
 		if err != nil {
@@ -147,17 +135,14 @@ func (s *UserService) DeleteUser(id int) error {
 	return s.userRepo.Delete(id)
 }
 
-// GetUserCount - количество пользователей
 func (s *UserService) GetUserCount() (int, error) {
 	return s.userRepo.Count()
 }
 
-// GetAdminCount - количество администраторов
 func (s *UserService) GetAdminCount() (int, error) {
 	return s.userRepo.CountAdmins()
 }
 
-// Authenticate - аутентификация пользователя
 func (s *UserService) Authenticate(username, password string) (*models.User, error) {
 	user, err := s.userRepo.GetByUsername(username)
 	if err != nil {
